requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-braces==1.14.0
django-filter==2.4.0
django-heroku==0.3.1
django-model-utils==4.1.0
django-registration==3.1.1
django-star==0.999
django-star-ratings==0.9.1
django-tinymce==3.1.0
django-vote==2.2.0
djangorestframework==3.12.2
greencache==1.0.0
gunicorn==20.0.4
htmq==1.0.0
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
redcmd==1.2.10
redlib==1.5.7
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stripe==2.55.1
swapper==1.1.2.post1
urllib3==1.26.2
vote==1.0.2
votes==1.0.1
whitenoise==5.2.0
zope.interface==5.2.0
